requirements.txt
asgiref==3.4.1
autopep8==1.5.7
certifi==2021.5.30
charset-normalizer==2.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.6
django-cors-headers==3.7.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2==2.9.1
pycodestyle==2.7.0
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
simplejson==3.17.4
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.6
whitenoise==5.3.0
